Update README to include prerequisite .NET 4.6.2 Developer Pack for windows
This fixes a bug with Lab05 when using a fresh install VS Community Edition 2017.

The `dotnet run -f net462` command would fail on Lab05 with a message
about needing the 4.6.2 framework. I reproduced the error on a fresh
Windows 7 install with VS2017 Community. Before installing, the error
occurred. After installing and opening a fresh command window, the
app runs successfully.